+++ CHANGELOG.md
@@ -13,6 +13,7 @@ CHANGED:
 
 - moved jobject-compare to its own project
 - moved jobect-evaluate to its own project
+- moved jobject-diff to its own project
 
 Version 2.6
 -------------