+++ CHANGELOG.md
@@ -4,9 +4,9 @@
 Version 3.1-SNAPSHOT
 -------------
 
-ADDED:
+CHANGED:
 
-- new methods for get all the declared fields of a class with an optional list of fields that shell be ignored
+- moved jobject-core to its own project
 
 Version 3
 -------------